requirements.txt
Flask
mysql-connector-python
flask-login